requirements.txt
absl-py==1.4.0
affine==2.4.0
aiofiles==23.2.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
astroid==3.0.1
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
autopep8==2.0.4
backcall==0.2.0
branca==0.7.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.6.5
decorator==4.4.2
dill==0.3.7
docstring-to-markdown==0.13
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
ExifRead==3.0.0
Faker==20.1.0
fastapi==0.112.0
ffmpy==0.4.0
filelock==3.13.1
fiona==1.9.5
fire==0.6.0
flake8==6.1.0
flatbuffers==23.5.26
folium==0.17.0
fonttools==4.43.1
fsspec==2023.12.0
gast==0.4.0
geopandas==0.14.0
google-auth==2.33.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.41.0
gradio_client==1.3.0
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib_resources==6.4.0
ipykernel==6.20.1
ipython==8.8.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==7.4.8
jupyter_core==5.1.3
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.2.1
numpy==1.24.3
oauthlib==3.2.2
onnx==1.16.0
onnxruntime==1.17.3
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.7
packaging==23.0
pandas==2.0.3
parso==0.8.3
pefile==2024.8.26
pickleshare==0.7.5
piexif==1.1.3
Pillow==10.0.0
platformdirs==4.0.0
pluggy==1.3.0
pocketsphinx==5.0.2
proglog==0.1.10
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydocstyle==6.3.0
pydub==0.25.1
pyexif==1.0.0
pyflakes==3.1.0
Pygments==2.14.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pylint==3.0.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
PyPDF2==3.0.1
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
python-multipart==0.0.9
pytoolconfig==1.2.6
pytz==2023.3
pywin32==305
pywin32-ctypes==0.2.3
PyYAML==6.0.1
pyzmq==24.0.1
rasterio==1.3.9
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rope==1.11.0
rsa==4.9
ruff==0.5.7
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
semantic-version==2.10.0
shapely==2.0.2
shellingham==1.5.4
shiboken6==6.6.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
snuggs==1.4.7
sounddevice==0.4.6
SpeechRecognition==3.10.0
stack-data==0.6.2
starlette==0.37.2
supervision==0.16.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.0
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1
tornado==6.2
tqdm==4.65.0
traitlets==5.8.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.8.0
ultralytics==8.0.222
ultralytics-yolov5==0.1.1
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.5
websockets==12.0
Werkzeug==2.3.6
whatthepatch==1.0.5
wrapt==1.15.0
xyzservices==2024.6.0
yapf==0.40.2
zipp==3.17.0
